Initial work toward efficient integer sets.
The two sets here (PosIntSet and IntSet) use a similar approach.
IntSet uses a second array of bytes to mark whether slots are used,
whereas PosIntSet is able to use non-positive numbers to signal this.

Benchmarks show that both approaches are faster than Miguel's previous
implementation. PosIntSet is fastest, but of course only supports a
limited set of inputs.

Currently PosIntSet#+= doesn't do any kind of error-checking to make
sure input is valid. This is trivial to add if necessary, but may not
be necessary if it is adapted to work with a "pre-validated" value
class.

Both classes extend Iterable[Int] and Function[Int => Boolean].