Added SpecializedSet.
The implementation is basically the same as for IntSet, but with
some changes to help specialization do the right thing. For instance,
we use the SpUnit[A] type to trick the compiler into specializing
the resize() method, even though we just return null.

The performance of SpecializedSet[Int] somehow managed to be better
than IntSet on some of the benchmarks, which I don't really understand.
Worth looking into.